+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v0.5.1
 
 - refactoring (interface manager)
+- adding formatter widgets